# api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package api

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// AuthResponse defines model for AuthResponse.
type AuthResponse struct {
	Token *string `json:"token,omitempty"`
}

// Comment defines model for Comment.
type Comment struct {
	Content   *string    `json:"content,omitempty"`
	CreatedAt *time.Time `json:"created_at,omitempty"`
	Id        *int       `json:"id,omitempty"`
	PostId    *int       `json:"post_id,omitempty"`
	UserId    *int       `json:"user_id,omitempty"`
}

// CommentCreate defines model for CommentCreate.
type CommentCreate struct {
	Content string `json:"content"`
	PostId  int    `json:"post_id"`
}

// CommentUpdate defines model for CommentUpdate.
type CommentUpdate struct {
	Content *string `json:"content,omitempty"`
}

// Post defines model for Post.
type Post struct {
	Content   *string    `json:"content,omitempty"`
	CreatedAt *time.Time `json:"created_at,omitempty"`
	Id        *int       `json:"id,omitempty"`
	Title     *string    `json:"title,omitempty"`
	UserId    *int       `json:"user_id,omitempty"`
}

// PostCreate defines model for PostCreate.
type PostCreate struct {
	Content string `json:"content"`
	Title   string `json:"title"`
}

// PostRating defines model for PostRating.
type PostRating struct {
	PostId *int     `json:"post_id,omitempty"`
	Rating *float32 `json:"rating,omitempty"`
}

// PostUpdate defines model for PostUpdate.
type PostUpdate struct {
	Content *string `json:"content,omitempty"`
	Title   *string `json:"title,omitempty"`
}

// User defines model for User.
type User struct {
	Id       *int    `json:"id,omitempty"`
	Username *string `json:"username,omitempty"`
}

// UserLogin defines model for UserLogin.
type UserLogin struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// UserRegister defines model for UserRegister.
type UserRegister struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// PostLoginJSONRequestBody defines body for PostLogin for application/json ContentType.
type PostLoginJSONRequestBody = UserLogin

// PostPostsJSONRequestBody defines body for PostPosts for application/json ContentType.
type PostPostsJSONRequestBody = PostCreate

// PatchPostsCommentCommentIdJSONRequestBody defines body for PatchPostsCommentCommentId for application/json ContentType.
type PatchPostsCommentCommentIdJSONRequestBody = CommentUpdate

// PostPostsCommentsJSONRequestBody defines body for PostPostsComments for application/json ContentType.
type PostPostsCommentsJSONRequestBody = CommentCreate

// PatchPostsPostIdJSONRequestBody defines body for PatchPostsPostId for application/json ContentType.
type PatchPostsPostIdJSONRequestBody = PostUpdate

// PostRegisterJSONRequestBody defines body for PostRegister for application/json ContentType.
type PostRegisterJSONRequestBody = UserRegister

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Login user
	// (POST /login)
	PostLogin(w http.ResponseWriter, r *http.Request)
	// Get list of posts
	// (GET /posts)
	GetPosts(w http.ResponseWriter, r *http.Request)
	// Create new post
	// (POST /posts)
	PostPosts(w http.ResponseWriter, r *http.Request)
	// Delete comment
	// (DELETE /posts/comment/{comment_id})
	DeletePostsCommentCommentId(w http.ResponseWriter, r *http.Request, commentId int)
	// Update comment
	// (PATCH /posts/comment/{comment_id})
	PatchPostsCommentCommentId(w http.ResponseWriter, r *http.Request, commentId int)
	// Get comments list
	// (GET /posts/comments)
	GetPostsComments(w http.ResponseWriter, r *http.Request)
	// Create comment
	// (POST /posts/comments)
	PostPostsComments(w http.ResponseWriter, r *http.Request)
	// Get comment by ID
	// (GET /posts/comments/{comment_id})
	GetPostsCommentsCommentId(w http.ResponseWriter, r *http.Request, commentId int)
	// Get post rating
	// (GET /posts/rating/{post_id})
	GetPostsRatingPostId(w http.ResponseWriter, r *http.Request, postId int)
	// Get posts by user ID
	// (GET /posts/users/{user_id})
	GetPostsUsersUserId(w http.ResponseWriter, r *http.Request, userId int)
	// Delete post
	// (DELETE /posts/{post_id})
	DeletePostsPostId(w http.ResponseWriter, r *http.Request, postId int)
	// Get post by ID
	// (GET /posts/{post_id})
	GetPostsPostId(w http.ResponseWriter, r *http.Request, postId int)
	// Update post
	// (PATCH /posts/{post_id})
	PatchPostsPostId(w http.ResponseWriter, r *http.Request, postId int)
	// Register new user
	// (POST /register)
	PostRegister(w http.ResponseWriter, r *http.Request)
	// Get list of users
	// (GET /users)
	GetUsers(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Login user
// (POST /login)
func (_ Unimplemented) PostLogin(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get list of posts
// (GET /posts)
func (_ Unimplemented) GetPosts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create new post
// (POST /posts)
func (_ Unimplemented) PostPosts(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete comment
// (DELETE /posts/comment/{comment_id})
func (_ Unimplemented) DeletePostsCommentCommentId(w http.ResponseWriter, r *http.Request, commentId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Update comment
// (PATCH /posts/comment/{comment_id})
func (_ Unimplemented) PatchPostsCommentCommentId(w http.ResponseWriter, r *http.Request, commentId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get comments list
// (GET /posts/comments)
func (_ Unimplemented) GetPostsComments(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create comment
// (POST /posts/comments)
func (_ Unimplemented) PostPostsComments(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get comment by ID
// (GET /posts/comments/{comment_id})
func (_ Unimplemented) GetPostsCommentsCommentId(w http.ResponseWriter, r *http.Request, commentId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get post rating
// (GET /posts/rating/{post_id})
func (_ Unimplemented) GetPostsRatingPostId(w http.ResponseWriter, r *http.Request, postId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get posts by user ID
// (GET /posts/users/{user_id})
func (_ Unimplemented) GetPostsUsersUserId(w http.ResponseWriter, r *http.Request, userId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete post
// (DELETE /posts/{post_id})
func (_ Unimplemented) DeletePostsPostId(w http.ResponseWriter, r *http.Request, postId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get post by ID
// (GET /posts/{post_id})
func (_ Unimplemented) GetPostsPostId(w http.ResponseWriter, r *http.Request, postId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Update post
// (PATCH /posts/{post_id})
func (_ Unimplemented) PatchPostsPostId(w http.ResponseWriter, r *http.Request, postId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Register new user
// (POST /register)
func (_ Unimplemented) PostRegister(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get list of users
// (GET /users)
func (_ Unimplemented) GetUsers(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PostLogin operation middleware
func (siw *ServerInterfaceWrapper) PostLogin(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostLogin(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetPosts operation middleware
func (siw *ServerInterfaceWrapper) GetPosts(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPosts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// PostPosts operation middleware
func (siw *ServerInterfaceWrapper) PostPosts(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPosts(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// DeletePostsCommentCommentId operation middleware
func (siw *ServerInterfaceWrapper) DeletePostsCommentCommentId(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "comment_id" -------------
	var commentId int

	err = runtime.BindStyledParameterWithOptions("simple", "comment_id", chi.URLParam(r, "comment_id"), &commentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "comment_id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeletePostsCommentCommentId(w, r, commentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// PatchPostsCommentCommentId operation middleware
func (siw *ServerInterfaceWrapper) PatchPostsCommentCommentId(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "comment_id" -------------
	var commentId int

	err = runtime.BindStyledParameterWithOptions("simple", "comment_id", chi.URLParam(r, "comment_id"), &commentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "comment_id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PatchPostsCommentCommentId(w, r, commentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetPostsComments operation middleware
func (siw *ServerInterfaceWrapper) GetPostsComments(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPostsComments(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// PostPostsComments operation middleware
func (siw *ServerInterfaceWrapper) PostPostsComments(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPostsComments(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetPostsCommentsCommentId operation middleware
func (siw *ServerInterfaceWrapper) GetPostsCommentsCommentId(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "comment_id" -------------
	var commentId int

	err = runtime.BindStyledParameterWithOptions("simple", "comment_id", chi.URLParam(r, "comment_id"), &commentId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "comment_id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPostsCommentsCommentId(w, r, commentId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetPostsRatingPostId operation middleware
func (siw *ServerInterfaceWrapper) GetPostsRatingPostId(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "post_id" -------------
	var postId int

	err = runtime.BindStyledParameterWithOptions("simple", "post_id", chi.URLParam(r, "post_id"), &postId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "post_id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPostsRatingPostId(w, r, postId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetPostsUsersUserId operation middleware
func (siw *ServerInterfaceWrapper) GetPostsUsersUserId(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "user_id" -------------
	var userId int

	err = runtime.BindStyledParameterWithOptions("simple", "user_id", chi.URLParam(r, "user_id"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "user_id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPostsUsersUserId(w, r, userId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// DeletePostsPostId operation middleware
func (siw *ServerInterfaceWrapper) DeletePostsPostId(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "post_id" -------------
	var postId int

	err = runtime.BindStyledParameterWithOptions("simple", "post_id", chi.URLParam(r, "post_id"), &postId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "post_id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeletePostsPostId(w, r, postId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetPostsPostId operation middleware
func (siw *ServerInterfaceWrapper) GetPostsPostId(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "post_id" -------------
	var postId int

	err = runtime.BindStyledParameterWithOptions("simple", "post_id", chi.URLParam(r, "post_id"), &postId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "post_id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPostsPostId(w, r, postId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// PatchPostsPostId operation middleware
func (siw *ServerInterfaceWrapper) PatchPostsPostId(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "post_id" -------------
	var postId int

	err = runtime.BindStyledParameterWithOptions("simple", "post_id", chi.URLParam(r, "post_id"), &postId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "post_id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PatchPostsPostId(w, r, postId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// PostRegister operation middleware
func (siw *ServerInterfaceWrapper) PostRegister(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostRegister(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetUsers operation middleware
func (siw *ServerInterfaceWrapper) GetUsers(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetUsers(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/login", wrapper.PostLogin)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/posts", wrapper.GetPosts)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/posts", wrapper.PostPosts)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/posts/comment/{comment_id}", wrapper.DeletePostsCommentCommentId)
	})
	r.Group(func(r chi.Router) {
		r.Patch(options.BaseURL+"/posts/comment/{comment_id}", wrapper.PatchPostsCommentCommentId)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/posts/comments", wrapper.GetPostsComments)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/posts/comments", wrapper.PostPostsComments)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/posts/comments/{comment_id}", wrapper.GetPostsCommentsCommentId)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/posts/rating/{post_id}", wrapper.GetPostsRatingPostId)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/posts/users/{user_id}", wrapper.GetPostsUsersUserId)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/posts/{post_id}", wrapper.DeletePostsPostId)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/posts/{post_id}", wrapper.GetPostsPostId)
	})
	r.Group(func(r chi.Router) {
		r.Patch(options.BaseURL+"/posts/{post_id}", wrapper.PatchPostsPostId)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/register", wrapper.PostRegister)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/users", wrapper.GetUsers)
	})

	return r
}